refactor: convert Similar Item function assistant to general function edit assistant
