+++ .github/workflows/on-pr-merge.yml
@@ -60,6 +60,7 @@ jobs:
 
       - name: Update and install system packages
         run: |
+          echo "VIRTUAL_ENV=ramstk-venvpy38" >> $GITHUB_ENV
           sudo apt-get update
           sudo apt-get install -y --no-install-recommends apt-utils git libgirepository1.0-dev gir1.2-gtk-3.0 python3-gi python3-gi-cairo python3-pytest libenchant1c2a libenchant-dev python3-numpy python3-scipy
 
@@ -69,12 +70,19 @@ jobs:
         with:
           python-version: 3.8
 
+      - name: Create a virtual environment
+        id: venv
+        run: |
+          python -mvenv $VIRTUAL_ENV
+
       - name: Install RAMSTK
         id: test
         run: |
+          source $VIRTUAL_ENV/bin/activate
           pip install -U pip setuptools>12.0 urllib3 poetry importlib_metadata
-          sudo make depends
-          sudo make install
+          python poetry_fix.py
+          make depends
+          make PREFIX=$VIRTUAL_ENV install
 
       - name: Build html documentation
         run: |
